feat: add country Components and refact code
