+++ rs-commerce/src/views/registration/components/RegistrationForm/components/AddressRegistration/components/AddressComponents/components/AddresTitleComponent/style.module.css
@@ -1,0 +1,14 @@
+.address__title-block {
+  display: flex;
+  justify-content: space-between;
+}
+
+.input__label {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.input_checkbox {
+  transform: scale(1.5);
+}
