+++ rs-commerce/src/views/registration/components/RegistrationForm/components/AddressRegistration/components/AddressComponents/style.module.css
@@ -1,14 +1,18 @@
-.address__title-block {
-  display: flex;
-  justify-content: space-between;
-}
-
 .address {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   grid-template-rows: repeat(2, 1fr);
   gap: 10px;
 }
+
+/* 
+
+.address__title-block {
+  display: flex;
+  justify-content: space-between;
+}
+
+
 
 .input__label {
   display: flex;
@@ -21,8 +25,13 @@
 }
 
 .checkbox__content {
-  color: rgb(0, 0, 0);
+  color: var(--dark);
   font-size: 14px;
   font-weight: 500;
   line-height: 20px;
 }
+
+.input_checkbox {
+  transform: scale(1.5);
+}
+ */
