+++ rs-commerce/src/views/registration/components/RegistrationForm/components/InputRegistration/CountryInput.tsx
@@ -1,15 +1,22 @@
+import { AddressType } from '../AddressRegistration/components/AddressComponents/components/AddresTitleComponent/AddresTitleComponent';
 import classes from './styles.module.css';
 
-function CountryInput(): JSX.Element {
+function CountryInput(props: AddressType): JSX.Element {
+  const { typeComponent } = props;
   return (
-    <label htmlFor="country" className={classes['fieldForm']}>
-      {' '}
-      Country:
-      <input id="country" list="countries" type="text" placeholder="Your country" />
-      <datalist id="countries">
-        <option>USA</option>
-        <option>Belarus</option>
-      </datalist>
+    <label htmlFor={`${typeComponent}Country`}>
+      <p className={classes['input__title']}>Country</p>
+      <div className={classes['input__block']}>
+        <input
+          id={`${typeComponent}Country`}
+          className={`${classes['registration__input']} ${classes['registration__input_small']}`}
+          list="country"
+        />
+        <datalist id="country">
+          <option>USA</option>
+          <option>Belarus</option>
+        </datalist>
+      </div>
     </label>
   );
 }
