+++ rs-commerce/src/views/registration/components/RegistrationForm/components/InputRegistration/styles.module.css
@@ -1,17 +1,3 @@
-.input__title {
-  color: rgb(0, 0, 0);
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 24px;
-}
-
-.input__error {
-  color: rgb(235, 87, 87);
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 20px;
-}
-
 .registration__input {
   height: 40px;
   width: 400px;
@@ -20,4 +6,18 @@
 
 .registration__input_small {
   width: 240px;
+}
+
+.input__title {
+  color: var(--dark);
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 24px;
+}
+
+.input__error {
+  color: var(--mistake);
+  font-size: 13px;
+  font-weight: 500;
+  line-height: 20px;
 }
