+++ rs-commerce/src/views/registration/style.module.css
@@ -18,86 +18,3 @@ h3 {
   background-color: var(--lilac-light);
   border-radius: 10px;
 }
-
-.input__label {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.input_checkbox {
-  transform: scale(1.5);
-}
-
-.address__title-block {
-  display: flex;
-  justify-content: space-between;
-}
-
-.address {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: repeat(2, 1fr);
-  gap: 10px;
-}
-
-.registration-form__wrapper {
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-}
-
-.registration__input_small {
-  width: 240px;
-}
-
-.registration__subtitle-block {
-  display: flex;
-  justify-content: space-between;
-}
-
-.registration__column-title {
-  color: var(--gray-middle);
-  font-size: 20px;
-  font-weight: 500;
-  line-height: 24px;
-}
-
-.checkbox__content {
-  color: var(--gray-middle);
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 20px;
-}
-
-.registration__button {
-  display: block;
-  margin: 20px auto 0;
-  width: 237px;
-  height: 55px;
-  background-color: var(--lilac-dark);
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  color: var(--white);
-}
-
-.input__title {
-  color: rgb(0, 0, 0);
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 24px;
-}
-
-.input__error {
-  color: rgb(235, 87, 87);
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 20px;
-}
-
-.registration__input {
-  height: 40px;
-  width: 400px;
-  font-size: 24px;
-}
